fix tooltip positioning
Add `position: relative` to make `#order-buttons div.tooltip` elements the
"closest positioned ancestor" of the order button tooltip `span` elements, so
that the `bottom` position is based on those tooltip divs.

Tooltips were initially positioned in [pull request #61][0] for work on
EVEREST-255, then that positioning was broken by seemingly unrelated CSS changes
in [pull request #67][1] for work on EVEREST-254. I took the CSS changes in this
commit and tested them back against those PR's merge commits, and it works in
both cases. In other words, if this had been how the positioning was originally
implemented in #61, it would not have been broken by #67.

[0]: https://bitbucket.org/nsidc/everest-ui/pull-requests/61/
[1]: https://bitbucket.org/nsidc/everest-ui/pull-requests/67/