Enable simultaneous dev in standalone and Drupal
In short: `npm run build-dev-drupal` now works for developing the app in Drupal
and in the standalone simultaneously by using `webpack-dev-server` along with a
plugin to write the files to the share location.

To test style changes that appear different in the standalone and in Drupal, it
should be possible to develop in both environments at once.

The plugin `write-file-webpack-plugin` forces `webpack-dev-server` to write the
files to the directory specified by the webpack config `output`. The GitHub
discussion about people wanting this feature in `webpack-dev-server` only to be
denied until one of them writes a plugin to do it was an entertaining read -
[webpack-dev-server #62][0].

When building/developing for the Drupal environment, `CESIUM_BASE_URL` is set to
`/apps/everest-ui`. This points to the location that webpack actually sees as
`/share/apps/everest-ui`. Therefore, `devServer.contentBase` is updated to
include `/share`.

[0]: https://github.com/webpack/webpack-dev-server/issues/62