ensure TemporalFilter is dealing with UTC dates
Fixes the bug where when a start date is loaded from localStorage, it is
displayed as the day before. This was because the date is actually a full
timestamp at midnight of the chosen date, and Moment infers the user's time zone
and then the timestamp was being displayed according to that time zone; since
MDT is UTC-6, this caused the date to be displayed as one day earlier. Adding
these `.utc()` calls ensures that Moment treats it only as a UTC date-time,
without respect for the user's time zone.